refactor: setting & config
