+++ crates/pm/src/cmd/view.rs
@@ -7,7 +7,7 @@ use utoo_ruborist::service::fetch_full_manifest;
 use utoo_ruborist::util::parse_package_spec;
 
 use crate::helper::ruborist_context::Context;
-use crate::util::config::get_registry;
+use crate::util::settings::get_registry;
 use crate::util::format_print::print_grid;
 
 /// View package information from registry, similar to npm view
