+++ crates/pm/src/helper/ruborist_context.rs
@@ -5,7 +5,7 @@ use utoo_ruborist::service::{BuildDepsOptions, Glob, UnifiedRegistry};
 
 use crate::service::pipeline::{PipelineChannels, PipelineReceiver};
 use crate::util::cache::get_cache_dir;
-use crate::util::config::{
+use crate::util::settings::{
     get_legacy_peer_deps, get_manifests_concurrency_limit, get_registry, get_supports_semver,
 };
 use crate::util::logger::ProgressReceiver;
