+++ crates/pm/src/service/binary.rs
@@ -1,5 +1,5 @@
 use crate::fs;
-use crate::util::config::get_registry;
+use crate::util::settings::get_registry;
 use crate::util::json::load_package_json_from_path;
 use anyhow::{Context, Result};
 use regex::Regex;
